Jagreda se realiza ajuste para la pestaña ANCLA
